+++ frontend/src/pages/Calendar/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const MainWrapper = styled.table`
+    width: 100vw;
+    height: calc(100vh - 125px);
+    min-width: 228px;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+`;
